task17899: Fixed buffer over-read bug during warmup.
During warmup we will ping-pong reply-sized messages.  The size of that can
be bigger than the test message size, and the buffer would only be the test
size.

Also during warmup prior to the RTT phase we would send messages of test
message size while receiver is expecting reply-sized messages.  No bad
effects on Linux, but this is causing a problem for port to Windows.  Also
this would cause problems when we support stream protocols.

hg-sha:4d2972a9e8f0