bug32601: Backed out changeset 2012-10-30T09:35:57Z!bhorgan@solarflare.com
This change is not needed, and I want to keep the sfnettest code base nice
and clean.

The discussion on the bug indicates that n_this_time is zero.  As far as I can
see that can only happen if cfg_miniter is zero.  Fix: Don't do that.

hg-sha:c32b7bc8b7b5