cleanup: Fix up the sfnettest-extension repository
The original repository was created by copying files from sfnettest rather
than cloning it.  This meant that we could no longer simply push from one
repo to another.  In this commit, I take the tip of the default and apply
all the changes that have been made to the extensions repo.