bug32208: Update sfnt-pingpong to use ONLOAD_MSG_WARM msgs
This patch allows us to do performance testing of ONLOAD_MSG_WARM msgs.
When the '--warm' flag is enabled, when the recvs timeout we will do a warm
send.  When using the '--warm' flag, you should also enable the '--spin-gap'
flag.  And while we are spin gapping, we will periodically do warm sends.