bug34074: Add support for templated sends to sfnt-pingpong
The implementation allocates the templated send off the critical path.

The cmdline allows you to specify what percentage of the payload should be
updated in the critical path so we can measure the performance with
different message sizes and update sizes.

When allocation fails with E2BIG, then the application will exit without any
errors.

If it fails for any other reason, then the application exits abnormally.