Bug 54882: Introduce warm-up to start of sfnt-pingpong
We were noticing that the first round of ping-pongs in a test run was
noticably higher than the next round (sometimes 50% plus). This effect was
more pronounced on some test rigs than others, but it was verifiable that if
the first two payloads were swapped then the inaccurate result remained the
first in the output and that if the iteration count was increased then the
effect became less pronounced, indicating that the distribution of the results
in the first round changed over time.

This patch adds a warm-up period that would appear to make the results
consistent.

hg-sha:4fa3e2518601