QAI-10345: Upgraded sfnt-pingpong from signed 32-bit int to 64-bit
Fixes issue where obtaining TSC values, measuring the start and stop
before and after a ping respectively, there were instances were the
delta between start and stop would be so large that, when casting
to 32-bit, this would cause an overflow. This would often be
represented in the results table as a negative, but could also be
represented as a positive if the delta is so large that it overflows
into the positive space.

The idea behind this change is to ensure that when large delays
are obtained, they are correctly reported in the results table,
and not as a result of overflow which may be difficult to assess
or detect especially if this overflows into the positive space.
An overflow in 64-bit space would indicate something is very wrong.