QAI-10345: Upgraded sfnt_stats from signed 32-bit int to 64-bit
As results now stores 64-bit values, the associated statistical
functions/struct values used must also be adapted to use 64-bit values.

This was mostly converting the types, but the sfnt_qsort_compare_int64
function uses a different comparison implementation to the 32-bit
version, as the qsort template expects an int return; the original
method of doing so simply subtracted the values and returned that, but
in 64-bit this needs to compare them directly and return a
representative int result as described in the comment.

Most changes here should just be a type change to 64-bit space in
order to accommodate the 64-bit upgrade. Though, it should be noted
that no changes were made to sfnt-stream which still should be using
the 32-bit versions left in the sfnt_stats source file.