QAI-10345: Removed result int cast
With results upgraded to signed 64-bit int, the cast to int that was
originally used is no longer needed as sfnt_tsc_nsec returns int64_t.

Though, whilst the app should now be able to store the results properly,
it does not alleviate the issue of these large delays being present in
the first place.