QAI-10345: Updated result print statements to use INT64 type
Updated the print statement in do_test which prints out each
row of the results table in the output to handle signed
64-bit ints, as stored in the upgraded results. This should
still be represented in the same format as before.

Also updated the print statement that prints out the table
headers to better match the printing style of other print
statements in the file. This should be more readable.