QAI-10345: Fix negative std-dev in sfnt-pingpong
The std-dev result for runs of sfnt-pingpong are sometimes negative.
This is caused, in the variance calculation, by 'sumsq' being a
signed 64-bit int, so when diff * diff is calculated this may
overflow in positive 64-bit space. Changing the type of 'sumsq'
and 'variance' to double prevents this (though may sacrifice
precision).